add map/group constants into constants.asm
map 18.3 and 7.14 are both "Route 10" ? Need to confirm that they are
both, in fact, "Route 10". Even if they both point to the same map, we
can call them 10a and 10b so that the map constants will work.

generate_map_constants and generate_map_constant_labels have been
updated to deal with left-over issues in various labels and obscure
characters.

original-commit-id: 8f2221aa703b3ed0d98003a055c65ea794144b64