redo the vba autoplayer
This rewrites the basic methods to use the new primitives for vba
automation. There are still a number of bugs in here.

Also, the level grinding recursion is broken because there's no method
used to detect whether or not a wild battle is starting.

Waiting for a battle to really end is currently broken.

There's a problem when interacting with radios because of the sound
effect check in text_wait.

There should be unit tests, but there aren't any.