write inject_asm_into_rom
This method injects asm straight into the ROM loaded in the emulator. It
does not overwrite the ROM on the file system.

This method is much slower than the wram version because it involves
copying memory multiples and copying the entire ROM into python and then
sending it back to the emulator.