gfx: Read image metadata from filenames where applicable.
png has some incompatibilities with 2bpp. Among them:
- 2bpp is tiled, and can be arranged arbitrarily
- 2bpp does not have associated palette data
- png must be rectangular
- png must have specific dimensions

The requirements for an image vary, and are unwritten. One conversion method will not work for all images.

To resolve these conflicts metadata can be attached to images and related data.
So as to not create additional files without changing the file content, the filename is used.

Usage: metadata is defined as pseudoextensions.
.w<width>
	- force width to <width> px
.h<height>
	- force height to <height> px
.t<num>
	- pad the image by <num> tiles so the resulting png can be rectangular
	- remove said tiles when converting back to 2bpp

Example:
	gfx/title/logo.w160.t4.{lz,2bpp,png}
	- width 160px
	- pad by 4 tiles